Separate odometry into its own namespace, print odometry to terminal
