+++ include/greenhat/api.h
@@ -2,3 +2,5 @@
 
 #include "greenhat/drive.h"
 #include "greenhat/selection.h"
+#include "greenhat/odom.h"
+#include "greenhat/purepursuit.h"
