+++ src/main.cpp
@@ -1,4 +1,5 @@
 #include "main.h"
+#include "greenhat/odom.h"
 #include "pros/imu.hpp"
 
 pros::Controller master(CONTROLLER_MASTER);
@@ -25,6 +26,8 @@ void opcontrol() {
 
 		arcade(master.get_analog(ANALOG_LEFT_Y) * (double)100 / 127,
 		       master.get_analog(ANALOG_RIGHT_X) * (double)100 / 127);
+        
+        printf("x: %lf y: %lf angle: %lf\n", odom::global_x, odom::global_y, odom::heading);
 
 		pros::delay(20);
 	}
